ai/server/requirements.txt
# Cherenkov AI Server Dependencies

# Cloud LLM
anthropic>=0.18.0

# API Server

# API Server
fastapi>=0.109.0
uvicorn>=0.27.0
python-multipart>=0.0.6

# Utilities
pydantic>=2.0.0
python-dotenv>=1.0.0

# --- Tool Integrations ---

# Email (Resend)
resend>=0.7.0

# Airtable
pyairtable>=2.2.0

# Google Calendar
google-auth>=2.27.0
google-auth-oauthlib>=1.2.0
google-api-python-client>=2.118.0

# Apple Calendar (CalDAV)
caldav>=1.3.0

